Improved flexibility
Converted hashTable into a class template to accomodate values of
arbitrary types rather than exclusively strings. Unfortunately,this
means that you now have to #include the .cpp file rather than the .h,
which is generally more common.

Also removed checks that prevented users from hashing empty values and
keys because I haven't found a reliable way to check whether a generic
type is empty and haven't found that it's very important anyway. At
first, I was concerned with having a table full of empty objects, but
if that's what the user wants, so be it.

Changed @length from unsigned int to size_t for redundant security.

Templating this class has been kind of a nightmare with the linker,
but here are some of the resources I used in case I want to revisit
this or if someone else would like to:
https://isocpp.org/wiki/faq/templates#templates-defn-vs-decl
https://www.ibm.com/docs/en/zos/2.2.0?topic=only-class-templates-c